feat: HomeAdapter 데이터 바인딩을 사용하여 MVVM 적용
